+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 ## Version 2.4.0.1
   ## Feature
   - Added Import/Export functionality
+  - Added Regular Expression (Match and No Match) Criteria
 - ## Bug
   - Changed the way log files are transferred.
 ## Version 2.4.0.0